Added edit functionality

+++ to-do-list/src/ListItem.js
@@ -2,7 +2,7 @@ import React, {useState} from "react";
 import "./listItem.css";
 
 export const ListItem = (props) => {
-    const {items, deleteItem} = props;
+    const {items, deleteItem, editItem, onSave, handleEditedInput, cancelEdit} = props;
     const [checked, setChecked] = useState([]);
 
     const handleCheck = (event, item) => {
@@ -21,12 +21,24 @@ export const ListItem = (props) => {
 
     return (
         <div>
-            {items.map((item, index) => 
-            <div key={index}>
-                <input type="checkbox" id={item} name={item} onChange={(event) => handleCheck(event,item)}></input>
-                <label htmlFor={item} className={isChecked(item)}>{item}</label>
-                <button onClick={() => deleteItem(item)}>Delete</button>
-            </div>)}
+            {items.map((item, index) => {
+                if (item.editing) {
+                    return <div key={index}>
+                        <form onSubmit={(e) => onSave(e, item)}>
+                            <input id='edit-input' value={item.name} type="text" onChange={(e) => handleEditedInput(e, item)}/>
+                            <button type="button" onClick={() => cancelEdit(item)}>Cancel</button>
+                            <button type="submit">Save</button>
+                        </form>
+                    </div>
+                } else {
+                    return <div key={index}>
+                        <input type="checkbox" id={item.name} name={item.name} onChange={(e) => handleCheck(e,item.name)}></input>
+                        <label htmlFor={item.name} className={isChecked(item.name)}>{item.name}</label>
+                        <button onClick={() => editItem(item)}>Edit</button>
+                        <button onClick={() => deleteItem(item)}>Delete</button>
+                    </div>
+                }
+            })}
         </div>
     );
 }
